+++ src/main/java/at/hannibal2/skyhanni/events/SkipTabListLineEvent.kt
@@ -1,6 +1,6 @@
 package at.hannibal2.skyhanni.events
 
-import at.hannibal2.skyhanni.features.gui.compacttablist.TabLine
+import at.hannibal2.skyhanni.features.misc.compacttablist.TabLine
 import net.minecraftforge.fml.common.eventhandler.Cancelable
 
 @Cancelable
